[#9|feature] Add navigation between lectures
As to #9 it is pretty inconvenient to navigate between lectures now:
user have to make 2 clicks and some scrolling to move to the
previous/next lecture of the course after reading the current one. This
commit solves described issue.

Links to next/previos lecture in category were added to posts.

Closes #9